Split up monster file
